# PR #27: Implement KISS Testing Strategy - 75% Less Complexity, Same Protection