--- a/tests/test_decision_regression.py
+++ b/tests/test_decision_regression.py
@@ -12,7 +12,9 @@ from bnb_trading.signals.decision import decide_long
 
 
 @pytest.fixture
-def decision_context(test_config, sample_daily_data, sample_weekly_data):
+def decision_context(
+    test_config, sample_daily_data, sample_weekly_data
+) -> DecisionContext:
     """Create DecisionContext for regression testing."""
     return DecisionContext(
         closed_daily_df=sample_daily_data,