… .claude/settings.local.json
… .pre-commit-config.yaml
… TESTING_STRATEGY_SIMPLIFIED.md
… tests/analysis/test_fibonacci.py